+++ lib/radar/application.rb
@@ -19,7 +19,7 @@ module Radar
     attr_reader :name
     attr_reader :creation_location
 
-    def_delegators :config, :reporters, :data_extensions, :matchers
+    def_delegators :config, :reporters, :data_extensions, :matchers, :filters
 
     # Looks up an application which was registered with the given name.
     #